% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engineer_data.R
\name{engineer_data}
\alias{engineer_data}
\title{engineer your data for a use case}
\usage{
engineer_data(data, use_case = "team_games", tg_odds)
}
\arguments{
\item{data}{data.frame data to be engineered}

\item{use_case}{chrwhich use case to engineer; currently just supports 'team_games'}

\item{tg_odds}{chr data.frame with odds}
}
\value{
engineered data
}
\description{
engineer your data for a use case
}
