% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engineer_new_data.R
\name{engineer_new_data}
\alias{engineer_new_data}
\title{engineer new data for forecasts}
\usage{
engineer_new_data(use_case = "team_games", schedule, odds, data)
}
\arguments{
\item{use_case}{chr}

\item{schedule}{data.frame}

\item{odds}{data.frame}

\item{data}{data.frame}
}
\value{
list
}
\description{
engineer new data for forecasts
}
