% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{read data for building models}
\usage{
get_data(data_path = NULL, filename_pattern = ".csv$", stitch_together = TRUE)
}
\arguments{
\item{data_path}{chr path to data}

\item{filename_pattern}{chr regex for filename patterns to get}

\item{stitch_together}{bool should all the files be combined as one}
}
\value{
data.frame if stitch_together is TRUE, else list
}
\description{
read data for building models
}
