% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rolling_window_list.R
\name{make_rolling_window_list}
\alias{make_rolling_window_list}
\title{make lists of rolling window functions}
\usage{
make_rolling_window_list(countdays = FALSE)
}
\arguments{
\item{countdays}{bool should windows only be counts of days}
}
\value{
list
}
\description{
make lists of rolling window functions
}
