% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_data.R
\name{preproc_data}
\alias{preproc_data}
\title{preprocess data for nba models}
\usage{
preproc_data(
  data_list,
  use_case = "team_games",
  na_perc = 0.1,
  highcorr_thresh = 0.9
)
}
\arguments{
\item{data_list}{list of data.frames}

\item{use_case}{chr which use case to go with; this should change to methods at some point}

\item{na_perc}{num proportion of a column's values that are NA for that column to be nuked}

\item{highcorr_thresh}{num threshold for pairwise correlations between columns, used in recipes::step_corr()}
}
\value{
list of data for modeling
}
\description{
preprocess data for nba models
}
