% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{split data for preprocessing and modeling}
\usage{
split_data(
  data,
  use_case = "team_games",
  tg_seas_type = "Regular Season",
  tg_mod_splits = sort(c(1, 5, 11, 23)),
  tg_testtrain_split_prop = 0.8
)
}
\arguments{
\item{data}{data.frame}

\item{use_case}{chr}

\item{tg_seas_type}{chr}

\item{tg_mod_splits}{num}

\item{tg_testtrain_split_prop}{num}
}
\value{
list
}
\description{
split data for preprocessing and modeling
}
