% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyup_data.R
\name{tidyup_data}
\alias{tidyup_data}
\title{tidy up your data for modeling}
\usage{
tidyup_data(
  data = NULL,
  fix_names = TRUE,
  fix_dates = TRUE,
  fix_numerics = TRUE,
  date_fields,
  numeric_fields
)
}
\arguments{
\item{data}{data.frame}

\item{fix_names}{bool}

\item{fix_dates}{bool}

\item{fix_numerics}{bool}

\item{date_fields}{chr fields to convert to dates in lowercase}

\item{numeric_fields}{chr fields to convert to numerics in lowercase}
}
\value{
data.frame
}
\description{
tidy up your data for modeling
}
